+++ GLLara/GLLVertexAttribAccessorSet.swift
@@ -26,13 +26,13 @@ import Foundation
     }
     
     func vertexFormat(vertexCount: Int, hasIndices: Bool) -> GLLVertexFormat {
-        return GLLVertexFormat(attributes: accessors.map { $0.attribute }, countOfVertices: UInt(vertexCount), hasIndices: hasIndices)
+        return GLLVertexFormat(attributes: accessors.map { $0.attribute }, countOfVertices: vertexCount, hasIndices: hasIndices)
     }
 }
 
 extension GLLVertexAttribAccessor {
     func typedElement<T>(at: Int, type: T.Type) -> UnsafePointer<T> {
         // TODO do a runtime check here
-        return element(at: UInt(at)).bindMemory(to: T.self, capacity: Int(attribute.numberOfElements))
+        return element(at: at).bindMemory(to: T.self, capacity: Int(attribute.numberOfElements))
     }
 }
